requirements.txt
astroid==2.3.3
Django==2.2.7
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
Pillow==6.2.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
torch==1.3.1+cpu
torchvision==0.4.2+cpu
typed-ast==1.4.0
wrapt==1.11.2
